web/requirements.txt
amqp==2.1.4
anyjson==0.3.3
beautifulsoup4==4.5.3
billiard==3.5.0.2
celery==4.0.2
Django==1.10.6
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-filter==1.0.1
django-model-utils==2.6.1
djangorestframework==3.6.2
docopt==0.6.2
drf-nested-routers==0.90.0
factory-boy==2.8.1
Faker==0.7.10
hues==0.2.2
instaLooter==0.8.0
itunes-iap==2.3.0
kombu==4.0.2
Markdown==2.6.8
munch==2.1.1
prettyexc==0.6.0
progressbar2==3.16.0
psycopg2==2.7.1
python-dateutil==2.6.0
python-utils==2.0.1
pytz==2016.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
vine==1.1.3
